panda add frontend, printer policy, Office Factory, web.php, User Model changes (Csak a Printer Policy megy a többi nem)

+++ app/Http/Controllers/PrinterController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Printer;
 use App\Http\Requests\StorePrinterRequest;
 use App\Http\Requests\UpdatePrinterRequest;
+use Illuminate\Support\Facades\Auth;
 
 class PrinterController extends Controller
 {
@@ -13,6 +14,9 @@ class PrinterController extends Controller
      */
     public function index()
     {
+        if (Auth::user()->cannot('view', Printer::class)) {
+            abort(403);
+        }
         $printers = Printer::all();
         return view('printers.index',['printers' => $printers]);
     }
@@ -22,7 +26,9 @@ class PrinterController extends Controller
      */
     public function create()
     {
-
+        if (Auth::user()->cannot('create', Printer::class)) {
+            abort(403);
+        }
         return view('printers.create');
     }
 
@@ -31,6 +37,9 @@ class PrinterController extends Controller
      */
     public function store(StorePrinterRequest $request)
     {
+        if (Auth::user()->cannot('create', Printer::class)) {
+            abort(403);
+        }
         $request->picture->storeAs(
             'picture',
             'printer_picture' . $request->brand . '_' . $request->type . '.jpg',
@@ -40,6 +49,8 @@ class PrinterController extends Controller
 
         $printer = Printer::create($request->all());
         $printer->picture = $file_name;
+        $printer->updated_at = now();
+        $printer->created_at = now();
         $printer->update();
 
         return redirect()->route("printers.index")->with("success", "Printer created successfully.");
@@ -50,6 +61,9 @@ class PrinterController extends Controller
      */
     public function show(Printer $printer)
     {
+        if (Auth::user()->cannot('viewAny', Printer::class)) {
+            abort(403);
+        }
         return view('printers.show', ['printer' => $printer]);
     }
 
@@ -58,7 +72,9 @@ class PrinterController extends Controller
      */
     public function edit(Printer $printer)
     {
-        //
+        if (Auth::user()->cannot('update', Printer::class)) {
+            abort(403);
+        }
         return view('printers.edit', ['printer' => $printer]);
     }
 
@@ -67,6 +83,9 @@ class PrinterController extends Controller
      */
     public function update(UpdatePrinterRequest $request, Printer $printer)
     {
+        if (Auth::user()->cannot('update', Printer::class)) {
+            abort(403);
+        }
         $printer->update($request->all());
         if($request->picture != null){
             $file_name = 'printer_picture' . $request->brand . '_' . $request->type . '.jpg';
@@ -76,6 +95,7 @@ class PrinterController extends Controller
                 'printer_picture' . $request->brand . '_' . $request->type . '.jpg',
                 'public');
         }
+        $printer->updated_at = now();
         $printer->update();
 
         return redirect()->route("printers.index")->with("success", "Printer updated successfully.");
@@ -86,17 +106,26 @@ class PrinterController extends Controller
      */
 
     public function updateUtilities(Printer $printer){
+        if (Auth::user()->cannot('updateUtilities', Printer::class)) {
+            abort(403);
+        }
         return view('printers.updateUtilities',['printer' => $printer]);
     }
 
     public function destroy(Printer $printer)
     {
+        if (Auth::user()->cannot('delete', Printer::class)) {
+            abort(403);
+        }
         $printer->delete();
         return back()->with('message', $printer->brand . ' ' . $printer->type . ' was deleted Successfully');
     }
 
     public function show_deleted()
     {
+        if (Auth::user()->cannot('view', Printer::class)) {
+            abort(403);
+        }
         $printers = Printer::withTrashed()->get();
         return view('printers.show_deleted',['printers' => $printers]);
     }
